+++ web/handler_test.go
@@ -11,6 +11,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:build go1.14 && !nobcrypt
+// +build go1.14,!nobcrypt
+
 package web
 
 import (
